Eyes are now googly
On top of that I really cleaned up the initialization code by creating World, Blob and Eye objects. The World is responsible for keeping track of all the constraints and points, the Blob takes care of initializing the system, and the Eyes keep track of their points and are responsible for rendering themselves.

To get the pupils to jiggle I set up a fixed width constraint between the pupil and the center of the eye. Whenever the pupil moves beyond its constraint's maximum, the constraint simply moves the pupil back inside. This give the illusion that the pupil is hitting the edge of the sclera (the white part of the eye).