Converted physics units to be independent of drawing units
The units used for physics are now entirely different from the units used for drawing. This gives me great flexibility in terms of tweaking both the physics and the drawing of things. As an added bonus the simulation now looks the same no matter what screen size.

Finally I added some simple debug drawing of the eyes that should hopefully help as I move ahead.