Adding in collisions between points and walls
Eyes are now aware of each other an will no longer overlap. I also finally hacked back in wall detection so the blob will not drift away. There isn't really any fancy physics to do that, it's more of a brute force approach that just checks if each point is overlapping with another.

I also documented all of the physics classes so far.