Rough pass at a computing a convex hull
It could probably be optimized further, but it does the job for the first attempt. I decided to go with a Monotone chain algorithm as it seemed like it had the least hairy math an is pretty quick to boot. The basics of the algorithm is to first sort the points from the top left to the bottom right. Next you repeatedly compute the cross product to find the outermost points.

I also optimized the World class to be a little smarter when checking interactions. Instead of blindly looping over all points, I now keep separate arrays for interactive and non-interactive points. This now allows me to limit the amount of looping inside of World.interact.