Expanding the convex hull outward
The hull should now be scaled outward such that it looks like all the googly eyes are actually contained within the hull. To do this I just compute the centre of the blob, then scale all the hull points out from there.

There is still a little distortion when the blobs are oblong in shape. This is caused by the fact I am using a single scale factor for X and Y. At some point it might be worth it to calculate the width to height ratio and create a scale factor for each dimension. In doing this I could probably get the padding around the eyes to be uniform in all directions.