Adding spline smoothing to the convex hull.
After a ton of research I finally through together some spline calculations for the hull of the blob. I'm not entirely happy with this implementation as it falls apart when two points of the hull are close together. I've already tried a couple of times to fix the jankiness with no success. I guess I'll just leave it as is for now. Maybe it will grow on me.