Eyes now follow the cursor
Each eyeball now gets it pupil's default force updated to point at the cursor. I also changed the interaction with the eyeballs to force the user to click directly on an eyeball to drag it. Lastly I increased the amount of default connections to make the hull slightly less jumpy. I'm still going to have to figure out a better plan for the hull jankiness.