Added a mouth, debug (x-ray) mode, and a reset button
The mouth is probably the crowning achievement of this commit. At several points while working on this I found myself chuckling out loud. Blobs now have the ability to feel:

* Happy if nothing if the user is moving their mouse.
* Content if nothing is going on.
* Worried if the user moves their mouse too close.
* Scared if the user grabs any member of the blob.
* Terror if the blob is being shaken.

On top of the above state machine, the mouth also does some simple tweening between states. This is done by keeping a separate list of points for what will actually be drawn, then with every render I simply check if the current points match the list of points of what emotion the mouth should be showing. If they don't match, I just increment or decrement the current points towards the specified emotion by some constant step.

The other awesome part of the commit is the debug button. I really like the idea of giving users the ability to see the inner workings of their blob. The only real achievement with debug mode is that I can now draw the spline control points. Everything else is pretty simple.

Lastly the reset button is what actually prompted most this work. It provided the opportunity the refactor the random code living in index.html into experiment.js. This is going to provide a great place to dump system level utilities and features.

At this point I am feeling great about this experiment and am super excited that there are still so many possibilities.