Added in gravity controls
There is now a new gravity control in the top left that provides some fun interaction. The user can start clicking and dragging on the button to control the direction and magnitude of the gravity. They can also simply click to toggle between the previously set gravity and no gravity.