Adding in device orientation and touch events
The blob should now work pretty well on most mobile devices. This finally gave me a chance to play around with the device motion/orientation APIs.

One interesting fact about those APIs is that they do not work so well when the device held vertically. Initially I had planned to make it such that if the user was holding their device vertically and rotated side to side that it would keep the gravity pointing down. Unfortunately I wasn't able to get this working reliably so I had to fall back on expecting the user to hold the device horizontally and tip it either front or back or side to side.