Added an about panel and making-of page
The making of page is the real jewel of this commit. The basic idea was to create a page where a user could see and interact with every version of the blob throughout its development.

The first step in doing this was writing a simple python script that loops over each commit and:

* Generates and unpacks a git archive for the given hash.
* Runs the build script at that hash (if it has one).
* Pulls out the relevant metadata (Subject, body, commit time, hash) and eventually dump it into a JS object that will be accessed later.

The second part involved building a simple page that loops over each commit in the aforementioned JS object and creating and iframe and a couple other elements for the metadata.

This was easy enough, leaning on John Resig's simple js templating library), but very quickly I discovered that running 10+ blob simulations simultaneously was not the best for performance… (read: my laptop sounded like it would catch fire at any moment).

To fix this I did two things.

* Do not even load the iframe until it has been scrolled into view.
* If an iframe has been loaded but is not currently visible then pause the simulation and animation.

Not loading the iframe until visible is trivial (just don't set the source), but pausing the simulations is where it got interesting. To do this I just reached into each non-visible iframe's contentWindow and replaced setTimeout, setInterval and requestAnimationFrame with no-op functions that store the last call's arguments. Then when an iframe is scrolled back into view I simply undo the no-op function and call the real native function again with its last call's arguments.

Overall I think the page performs really well now and the user doesn't even know that non-visible blobs are paused.